+++ src/components/Pagination/Pagination.story.js
@@ -6,11 +6,18 @@ import { withKnobs, number } from '@storybook/addon-knobs';
 
 import Pagination from './index.js';
 
-const stories = storiesOf('Navigation|Pagination', module);
+const stories = storiesOf('Pagination (deprecated)', module);
 stories
   .addDecorator(withKnobs)
   .addDecorator(story => (
     <div id="root-preview">
+      <p>
+        Please use
+        <a href="/?path=/story/navigation-paginationcontrols-using-buttons-no-href--default">
+          {' '}PaginationControls{' '}
+        </a>
+        component instead
+      </p>
       {story()}
     </div>
   ));