+++ src/components/SummaryBar/SummaryBar.story.tsx
@@ -5,7 +5,7 @@ import { withKnobs } from '@storybook/addon-knobs';
 
 import SummaryBar from './index';
 
-const stories = storiesOf('Layout|SummaryBar', module);
+const stories = storiesOf('Data Organization|SummaryBar', module);
 stories.addDecorator(withKnobs).addDecorator(story => <div id="root-preview">{story()}</div>);
 
 stories.add('Bar with mixed content', () => {